requirements.txt
asgiref==3.7.2
certifi==2022.9.24
charset-normalizer==3.0.1
configparser==5.3.0
contourpy==1.0.5
cycler==0.11.0
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
fonttools==4.37.3
gunicorn==21.2.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.0
mrjob==0.7.4
numpy==1.23.3
oauthlib==3.2.2
orjson==3.8.6
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
psycopg2==2.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2022.6
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sqlparse==0.4.4
tweepy==4.12.1
typing_extensions==4.9.0
urllib3==1.26.14
weepy==1.2.0
whitenoise==6.6.0
